refactor: unify store_init_kwargs code
