+++ jina/serve/executors/__init__.py
@@ -13,12 +13,9 @@ from jina.helper import ArgNamespace, T, iscoroutinefunction, typename
 from jina.importer import ImportExtensions
 from jina.jaml import JAML, JAMLCompatible, env_var_regex, internal_var_regex
 from jina.logging.logger import JinaLogger
-from jina.serve.executors.decorators import (
-    avoid_concurrent_lock_cls,
-    requests,
-    store_init_kwargs,
-)
-from jina.serve.helper import wrap_func
+from jina.serve.executors.decorators import avoid_concurrent_lock_cls, requests
+from jina.serve.executors.metas import get_default_metas
+from jina.serve.helper import store_init_kwargs, wrap_func
 
 if TYPE_CHECKING:
     from prometheus_client import Summary
@@ -63,7 +60,11 @@ class ExecutorType(type(JAMLCompatible), type):
                     f'{cls.__init__} does not follow the full signature of `Executor.__init__`, '
                     f'please add `**kwargs` to your __init__ function'
                 )
-            wrap_func(cls, ['__init__'], store_init_kwargs)
+            taboo = {'self', 'args', 'kwargs', 'metas', 'requests', 'runtime_args'}
+            _defaults = get_default_metas()
+            taboo.update(_defaults.keys())
+
+            wrap_func(cls, ['__init__'], store_init_kwargs, taboo=taboo)
             wrap_func(cls, ['__init__'], avoid_concurrent_lock_cls(cls))
 
             reg_cls_set.add(cls_id)
