+++ jina/serve/executors/decorators.py
@@ -25,43 +25,6 @@ def _get_locks_root() -> Path:
     return locks_root
 
 
-def store_init_kwargs(func: Callable) -> Callable:
-    """Mark the args and kwargs of :func:`__init__` later to be stored via :func:`save_config` in YAML
-    :param func: the function to decorate
-    :return: the wrapped function
-    """
-
-    @functools.wraps(func)
-    def arg_wrapper(self, *args, **kwargs):
-        if func.__name__ != '__init__':
-            raise TypeError(
-                'this decorator should only be used on __init__ method of an executor'
-            )
-        taboo = {'self', 'args', 'kwargs', 'metas', 'requests', 'runtime_args'}
-        _defaults = get_default_metas()
-        taboo.update(_defaults.keys())
-        all_pars = inspect.signature(func).parameters
-        tmp = {k: v.default for k, v in all_pars.items() if k not in taboo}
-        tmp_list = [k for k in all_pars.keys() if k not in taboo]
-        # set args by aligning tmp_list with arg values
-        for k, v in zip(tmp_list, args):
-            tmp[k] = v
-        # set kwargs
-        for k, v in kwargs.items():
-            if k in tmp:
-                tmp[k] = v
-
-        if hasattr(self, '_init_kwargs_dict'):
-            self._init_kwargs_dict.update(tmp)
-        else:
-            self._init_kwargs_dict = tmp
-        convert_tuple_to_list(self._init_kwargs_dict)
-        f = func(self, *args, **kwargs)
-        return f
-
-    return arg_wrapper
-
-
 def avoid_concurrent_lock_cls(cls):
     """Wraps a function to lock a filelock for concurrent access with the name of the class to which it applies, to avoid deadlocks
     :param cls: the class to which is applied, only when the class corresponds to the instance type, this filelock will apply
